+++ CHANGES.md
@@ -1,10 +1,11 @@
 VECTO-CSE: Changes
 ===================
-#### 2016-11-25: v3.0.1 ####
+#### 2016-11-25: v3.0.1/2 ####
 TUG improvements:
 
   * New version number
   * Bugfix for digit check by use of GPS
+  * Bugfix for measurement section detection for root point calculation
   
 #### 2016-11-16: v3.0.0 ####
 TUG improvements: